+++ internal/ceres/covariance_test.cc
@@ -234,7 +234,7 @@ TEST(CovarianceImpl, ComputeCovarianceSparsity) {
   }
 
   const int* cols = crsm->cols();
-  for (int c = 0; c < crsm->num_nonzeros(); ++c) {
+  for (int64_t c = 0; c < crsm->num_nonzeros(); ++c) {
     EXPECT_EQ(cols[c], expected_cols[c])
         << c << " "
         << cols[c] << " "
@@ -317,7 +317,7 @@ TEST(CovarianceImpl, ComputeCovarianceSparsityWithConstantParameterBlock) {
   }
 
   const int* cols = crsm->cols();
-  for (int c = 0; c < crsm->num_nonzeros(); ++c) {
+  for (int64_t c = 0; c < crsm->num_nonzeros(); ++c) {
     EXPECT_EQ(cols[c], expected_cols[c])
         << c << " "
         << cols[c] << " "
@@ -398,7 +398,7 @@ TEST(CovarianceImpl, ComputeCovarianceSparsityWithFreeParameterBlock) {
   }
 
   const int* cols = crsm->cols();
-  for (int c = 0; c < crsm->num_nonzeros(); ++c) {
+  for (int64_t c = 0; c < crsm->num_nonzeros(); ++c) {
     EXPECT_EQ(cols[c], expected_cols[c])
         << c << " "
         << cols[c] << " "
